Split Helix fees adapter into spot and perp
Addresses review feedback. Replaces fees/helix.ts (combined) with two adapters mirroring the volume split: fees/helix-helix.ts (spot) and fees/helix-helix-perp.ts (perp). Each hits its dedicated Cloud Run endpoint (/helix_spot_fees and /helix_derivative_fees).
